refactor: Rename package name
