+++ src/goapp/repository/externallink/interface.go
@@ -1,4 +1,4 @@
-package repository
+package externallink
 
 import "main/model"
