+++ src/goapp/service/externallink/interface.go
@@ -1,4 +1,4 @@
-package service
+package externallink
 
 import "main/model"
